+++ src/Symfony/Component/Intl/ResourceBundle/Reader/BundleEntryReaderInterface.php
@@ -44,8 +44,9 @@ interface BundleEntryReaderInterface extends BundleReaderInterface
      *                           in the requested locale.
      *
      * @return mixed Returns an array or {@link \ArrayAccess} instance for
-     *               complex data, a scalar value for simple data and NULL
-     *               if the given path could not be accessed.
+     *               complex data and a scalar value for simple data.
+     *
+     * @throws \Symfony\Component\Intl\Exception\MissingResourceException If the indices cannot be accessed
      */
     public function readEntry($path, $locale, array $indices, $fallback = true);
 }